requirements2.txt
JPype1==0.7.0
lxml==4.1.0
numpy==1.16.6
